feat(users): New User.product_count field (#441)

+++ open_prices/users/tests.py
@@ -30,23 +30,25 @@ class UserPropertyTest(TestCase):
         cls.user = UserFactory()
         cls.location = LocationFactory(**LOCATION_NODE_652825274)
         PriceFactory(
-            owner=cls.user.user_id,
+            product_code="0123456789100",
             location_osm_id=cls.location.osm_id,
             location_osm_type=cls.location.osm_type,
             price=1.0,
+            owner=cls.user.user_id,
         )
         PriceFactory(
-            owner=cls.user.user_id,
+            product_code="0123456789101",
             location_osm_id=cls.location.osm_id,
             location_osm_type=cls.location.osm_type,
             price=2.0,
+            owner=cls.user.user_id,
         )
 
     def test_update_price_count(self):
         self.user.refresh_from_db()
         self.assertEqual(self.user.price_count, 2)
         # bulk delete prices to skip signals
-        Price.objects.filter(owner=self.user).delete()
+        Price.objects.filter(owner=self.user.user_id).delete()
         self.assertEqual(self.user.price_count, 2)  # should be 0
         # update_price_count() should fix price_count
         self.user.update_price_count()
@@ -58,3 +60,10 @@ class UserPropertyTest(TestCase):
         # update_location_count() should fix location_count
         self.user.update_location_count()
         self.assertEqual(self.user.location_count, 1)
+
+    def test_update_product_count(self):
+        self.user.refresh_from_db()
+        self.assertEqual(self.user.product_count, 0)
+        # update_product_count() should fix product_count
+        self.user.update_product_count()
+        self.assertEqual(self.user.product_count, 2)
